+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/init/PacketInitializer.java
@@ -2,6 +2,7 @@ package com.codetaylor.mc.pyrotech.modules.tech.basic.init;
 
 import com.codetaylor.mc.athenaeum.network.IPacketRegistry;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.network.SCPacketCapabilitySyncFocused;
+import com.codetaylor.mc.pyrotech.modules.tech.basic.network.SCPacketMarshmallowStickTimestamp;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.network.SCPacketParticleAnvilHit;
 import net.minecraftforge.fml.relauncher.Side;
 
@@ -18,6 +19,12 @@ public final class PacketInitializer {
     registry.register(
         SCPacketCapabilitySyncFocused.class,
         SCPacketCapabilitySyncFocused.class,
+        Side.CLIENT
+    );
+
+    registry.register(
+        SCPacketMarshmallowStickTimestamp.class,
+        SCPacketMarshmallowStickTimestamp.class,
         Side.CLIENT
     );
   }