+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/item/ItemMarshmallowStick.java
@@ -2,10 +2,15 @@ package com.codetaylor.mc.pyrotech.modules.tech.basic.item;
 
 import com.codetaylor.mc.athenaeum.util.StackHelper;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.ModuleTechBasic;
+import com.codetaylor.mc.pyrotech.modules.tech.basic.ModuleTechBasicConfig;
+import com.codetaylor.mc.pyrotech.modules.tech.basic.network.SCPacketMarshmallowStickTimestamp;
+import com.codetaylor.mc.pyrotech.modules.tech.basic.tile.TileCampfire;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
+import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.EntityEquipmentSlot;
+import net.minecraft.item.EnumAction;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
@@ -13,9 +18,13 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.World;
+import org.apache.logging.log4j.LogManager;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.Arrays;
 
 public class ItemMarshmallowStick
@@ -48,12 +57,48 @@ public class ItemMarshmallowStick
   public ItemMarshmallowStick() {
 
     this.setMaxStackSize(1);
-    this.setMaxDamage(8);
+    this.setMaxDamage(ModuleTechBasicConfig.CAMPFIRE_MARSHMALLOWS.MARSHMALLOW_STICK_DURABILITY);
 
     this.addPropertyOverride(
         new ResourceLocation(ModuleTechBasic.MOD_ID, "marshmallow_type"),
         (itemStack, world, entity) -> ItemMarshmallowStick.getType(itemStack).id
     );
+  }
+
+  // ---------------------------------------------------------------------------
+  // - Accessors
+  // ---------------------------------------------------------------------------
+
+  @Override
+  public int getMaxDamage(ItemStack stack) {
+
+    return ModuleTechBasicConfig.CAMPFIRE_MARSHMALLOWS.MARSHMALLOW_STICK_DURABILITY;
+  }
+
+  @Override
+  public int getMaxItemUseDuration(ItemStack stack) {
+
+    /*if (ItemMarshmallowStick.getType(stack) == EnumType.MARSHMALLOW) {
+      return 5 * 20;
+    }
+
+    return 0;*/
+    return 10 * 60 * 20;
+  }
+
+  @Nonnull
+  @Override
+  public EnumAction getItemUseAction(ItemStack stack) {
+
+    switch (ItemMarshmallowStick.getType(stack)) {
+
+      case MARSHMALLOW:
+      case MARSHMALLOW_ROASTED:
+        return EnumAction.BOW;
+
+      default:
+        return EnumAction.NONE;
+    }
   }
 
   @Nonnull
@@ -75,7 +120,43 @@ public class ItemMarshmallowStick
     return super.getUnlocalizedName(stack);
   }
 
+  @Override
+  public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+
+    return ItemMarshmallowStick.getType(oldStack) != ItemMarshmallowStick.getType(newStack);
+  }
+
+  @Nullable
+  @Override
+  public NBTTagCompound getNBTShareTag(ItemStack stack) {
+
+    // We remove the timestamp here because when it is sent to the client,
+    // it replaces the item on the client, and it will interrupt the item
+    // being used. We send a custom packet to the client and update this
+    // value manually to bypass the interrupt.
+
+    NBTTagCompound tag = super.getNBTShareTag(stack);
+
+    if (tag != null) {
+
+      // Allow the timestamp tag to be sent to the client if we're clearing
+      // the value, causing an item change.
+      //if (ItemMarshmallowStick.getRoastTimestamp(stack) != Long.MAX_VALUE) {
+      tag = tag.copy();
+      tag.removeTag("RoastByTimestamp");
+      //}
+    }
+    return tag;
+  }
+
+  // ---------------------------------------------------------------------------
+  // - NBT Accessors
+  // ---------------------------------------------------------------------------
+
   public static void setType(EnumType type, ItemStack itemStack) {
+
+    // TODO: remove loggers and stdout
+    LogManager.getLogger(ItemMarshmallowStick.class).info(type);
 
     NBTTagCompound tag = StackHelper.getTagSafe(itemStack);
     tag.setInteger("MarshmallowType", type.id);
@@ -88,15 +169,80 @@ public class ItemMarshmallowStick
     return EnumType.from(tag.getInteger("MarshmallowType"));
   }
 
+  public static void setRoastByTimestamp(World world, EntityPlayer player, ItemStack itemStack, long timestamp) {
+
+    // TODO: remove logger, unused params
+    LogManager.getLogger(ItemMarshmallowStick.class).info(timestamp);
+
+    NBTTagCompound tag = StackHelper.getTagSafe(itemStack);
+    tag.setLong("RoastByTimestamp", timestamp);
+    itemStack.setTagCompound(tag);
+  }
+
+  public static long getRoastByTimestamp(ItemStack itemStack) {
+
+    NBTTagCompound tag = StackHelper.getTagSafe(itemStack);
+
+    if (!tag.hasKey("RoastByTimestamp")) {
+      return Long.MAX_VALUE;
+    }
+
+    return tag.getLong("RoastByTimestamp");
+  }
+
+  // ---------------------------------------------------------------------------
+  // - Interaction
+  // ---------------------------------------------------------------------------
+
   @Nonnull
   @Override
   public ActionResult<ItemStack> onItemRightClick(World world, EntityPlayer player, @Nonnull EnumHand hand) {
+
+//    System.out.println("CLICK");
 
     if (hand == EnumHand.MAIN_HAND) {
       ItemStack itemMainHand = player.getHeldItemMainhand();
       ItemStack itemOffhand = player.getHeldItemOffhand();
 
+      // Try starting the roast.
+
+      if (ItemMarshmallowStick.getType(itemMainHand) != EnumType.EMPTY
+          /*|| itemOffhand.getItem() != ModuleTechBasic.Items.MARSHMALLOW*/) {
+
+        RayTraceResult rayTraceResult = this.rayTrace(world, player, false);
+
+        // ray trace result can be null
+        //noinspection ConstantConditions
+        if (rayTraceResult != null
+            && rayTraceResult.typeOfHit == RayTraceResult.Type.BLOCK
+            && this.isRoastingBlock(world, rayTraceResult.getBlockPos())) {
+
+          // TODO: Setting the NBT data here is causing an item reset and interrupting the use cycle.
+          // When we set the NBT later, set it to roasted, it will interrupt the roasting process and
+          // never let the marshmallow burn.
+          if (!world.isRemote && ItemMarshmallowStick.getRoastByTimestamp(itemMainHand) == Long.MAX_VALUE) {
+
+            // TODO: timestamp, config these magic numbers
+
+            int roastDuration = 5 * 20;
+            float roastVariance = 0.2f;
+            roastDuration = Math.max(0, (int) (roastDuration - roastDuration * roastVariance));
+
+            long timestamp = world.getTotalWorldTime() + roastDuration;
+            ItemMarshmallowStick.setRoastByTimestamp(world, player, itemMainHand, timestamp);
+            ModuleTechBasic.PACKET_SERVICE.sendToAll(new SCPacketMarshmallowStickTimestamp(player, timestamp));
+          }
+
+          player.setActiveHand(hand);
+          return new ActionResult<>(EnumActionResult.SUCCESS, itemMainHand);
+        }
+      }
+
+      // Try placing a marshmallow on the stick or removing one from the stick.
+
       if (itemOffhand.getItem() == ModuleTechBasic.Items.MARSHMALLOW) {
+
+        // Place a marshmallow on the stick.
 
         switch (ItemMarshmallowStick.getType(itemMainHand)) {
           case MARSHMALLOW_BURNED:
@@ -122,9 +268,14 @@ public class ItemMarshmallowStick
 
       } else if (itemOffhand.isEmpty()) {
 
+        // Remove a marshmallow from the stick.
+
         switch (ItemMarshmallowStick.getType(itemMainHand)) {
           case MARSHMALLOW_BURNED:
-            // TODO: remove burned marshmallow
+            player.setItemStackToSlot(EntityEquipmentSlot.OFFHAND, new ItemStack(ModuleTechBasic.Items.MARSHMALLOW_BURNED));
+            ItemMarshmallowStick.setType(EnumType.EMPTY, itemMainHand);
+            this.setCooldownOnMarshmallows(player);
+            return new ActionResult<>(EnumActionResult.SUCCESS, itemMainHand);
 
           case MARSHMALLOW_ROASTED:
             player.setItemStackToSlot(EntityEquipmentSlot.OFFHAND, new ItemStack(ModuleTechBasic.Items.MARSHMALLOW_ROASTED));
@@ -147,10 +298,111 @@ public class ItemMarshmallowStick
     return super.onItemRightClick(world, player, hand);
   }
 
+  @Override
+  public void onUsingTick(ItemStack stack, EntityLivingBase player, int count) {
+
+//    System.out.println("USING");
+
+    if (!(player instanceof EntityPlayer)) {
+      return;
+    }
+
+    World world = player.world;
+    RayTraceResult rayTraceResult = this.rayTrace(world, (EntityPlayer) player, false);
+
+    // The ray trace result can be null
+    //noinspection ConstantConditions
+    if (rayTraceResult == null
+        || rayTraceResult.typeOfHit != RayTraceResult.Type.BLOCK
+        || !this.isRoastingBlock(world, rayTraceResult.getBlockPos())) {
+
+      // TODO: Check if within roasting range of the fire
+
+      player.stopActiveHand();
+      this.setCooldownOnMarshmallows((EntityPlayer) player);
+    }
+
+    // TODO: turn into a roasted stick when it passes the threshold
+    if (world.isRemote) {
+      long totalWorldTime = world.getTotalWorldTime();
+      long roastTimestamp = ItemMarshmallowStick.getRoastByTimestamp(stack);
+
+      if (roastTimestamp < Long.MAX_VALUE) {
+        EnumType type = ItemMarshmallowStick.getType(stack);
+
+        if (type == EnumType.MARSHMALLOW_ROASTED && totalWorldTime >= roastTimestamp + 20) { // TODO: magic numbers
+          ItemMarshmallowStick.setType(EnumType.MARSHMALLOW_BURNED, stack);
+
+        } else if (type == EnumType.MARSHMALLOW && totalWorldTime >= roastTimestamp) {
+          ItemMarshmallowStick.setType(EnumType.MARSHMALLOW_ROASTED, stack);
+        }
+      }
+    }
+  }
+
+  @Override
+  public void onPlayerStoppedUsing(ItemStack stack, World world, EntityLivingBase player, int timeLeft) {
+
+    System.out.println("STOPPED");
+
+    if (!(player instanceof EntityPlayer)) {
+      return;
+    }
+
+    long totalWorldTime = world.getTotalWorldTime();
+    long roastTimestamp = ItemMarshmallowStick.getRoastByTimestamp(stack);
+
+    if (totalWorldTime >= roastTimestamp + 20) { // TODO: magic numbers
+      ItemMarshmallowStick.setType(EnumType.MARSHMALLOW_BURNED, stack);
+
+    } else if (totalWorldTime >= roastTimestamp) {
+      ItemMarshmallowStick.setType(EnumType.MARSHMALLOW_ROASTED, stack);
+    }
+
+    ItemMarshmallowStick.setRoastByTimestamp(world, (EntityPlayer) player, stack, Long.MAX_VALUE);
+  }
+
+  @Nonnull
+  @Override
+  public ItemStack onItemUseFinish(@Nonnull ItemStack stack, World world, EntityLivingBase player) {
+
+    System.out.println("FINISH");
+
+    if (!(player instanceof EntityPlayer)) {
+      return stack;
+    }
+
+    ItemMarshmallowStick.setRoastByTimestamp(world, (EntityPlayer) player, stack, Long.MAX_VALUE);
+    ItemMarshmallowStick.setType(EnumType.MARSHMALLOW_BURNED, stack);
+
+    RayTraceResult rayTraceResult = this.rayTrace(world, (EntityPlayer) player, false);
+
+    // The ray trace result can be null
+    //noinspection ConstantConditions
+    if (rayTraceResult == null
+        || rayTraceResult.typeOfHit != RayTraceResult.Type.BLOCK
+        || !this.isRoastingBlock(world, rayTraceResult.getBlockPos())) {
+
+      // TODO: Check if within roasting range of the fire
+
+      player.stopActiveHand();
+      return stack;
+    }
+
+    // TODO: turn into a burned stick
+    return stack;
+  }
+
+  public boolean isRoastingBlock(World world, BlockPos blockPos) {
+
+    return world.getTileEntity(blockPos) instanceof TileCampfire;
+  }
+
   private void setCooldownOnMarshmallows(EntityPlayer player) {
 
     player.getCooldownTracker().setCooldown(this, 10);
     player.getCooldownTracker().setCooldown(ModuleTechBasic.Items.MARSHMALLOW, 10);
     player.getCooldownTracker().setCooldown(ModuleTechBasic.Items.MARSHMALLOW_ROASTED, 10);
+    player.getCooldownTracker().setCooldown(ModuleTechBasic.Items.MARSHMALLOW_BURNED, 10);
   }
 }